BUG: incorrect cast when reading scalar attributes

+++ src/postProcessing/functionObjects/IOadiosWrite/adiosWrite/adiosReader.C
@@ -389,7 +389,7 @@ bool Foam::adiosReader::readScalarAttributeIfPresent
 
             if (ok)
             {
-                value = *(reinterpret_cast<int*>(data));
+                value = *(reinterpret_cast<double*>(data));
             }
         }
         else
